separando as rotas de cada parte em sua proprio arquivo
