+++ src/projeto-backend/models/evidence_crud.py
@@ -5,6 +5,7 @@ from sqlalchemy import  Column, ForeignKey, Integer, String, Date
 
 
 class Evidence(Base):
+    """Classe para trabalhar a tabela de evidencia do banco"""
     __tablename__ = 'evidence'
 
     id = Column(Integer, primary_key=True, index=True, autoincrement=True)
@@ -16,13 +17,16 @@ class Evidence(Base):
 
 
 def get_evidence(db: Session, id: int):
+    """Busca a evidencia pelo id dela e retorna ela"""
     return db.query(Evidence).filter(Evidence.id==id).first()
 
 
 def get_all_evidence(db: Session, skip: int = 0, limit: int = 100):
+    """Retorna todas as evidencia do banco mas somente as 100 primeiras podendo ser paginado outras"""
     return db.query(Evidence).offset(skip).limit(limit).all()
 
 def create_evidence(db: Session, evidence: schemas.EvidenceCreate):
+    """Cria uma nova evidencia no banco"""
     db_evidence = Evidence(link=evidence.link, attachment=evidence.attachment, idRequestForEvidence=evidence.idRequestForEvidence, deliveryDate=evidence.deliveryDate)
     db.add(db_evidence)
     db.commit()
@@ -31,6 +35,7 @@ def create_evidence(db: Session, evidence: schemas.EvidenceCreate):
 
 
 def update_evidence(db: Session, evidence: schemas.Evidence):
+    """Se exister a evidencia no banco realiza a alteração dela e salva as alterações"""
     db_evidence = db.query(Evidence).filter(Evidence.id == evidence.id).first()
 
     if db_evidence:
@@ -45,6 +50,7 @@ def update_evidence(db: Session, evidence: schemas.Evidence):
     return db_evidence
 
 def delete_evidence(db: Session, id: int):
+    """Se a evidencia existir no banco é realizado a exclusão dela no banco pelo id dela"""
     db_evidence = db.query(Evidence).filter(Evidence.id == id).first()
 
     if db_evidence:
