+++ src/projeto-backend/routers/request_for_evidence.py
@@ -1,0 +1,35 @@
+
+from fastapi import APIRouter, Depends, HTTPException
+from database import schemas
+from sqlalchemy.orm import Session
+from models import request_for_evidence_crud
+from database.database import get_db
+
+router = APIRouter(
+    tags=['Request_For_Evidence']
+
+)
+
+
+
+## Pedidos De Evidencia rotas
+
+@router.get("/request_for_evidence/{id}", response_model=schemas.RequestForEvidence)
+def get_request_for_evidence(id: int, db: Session = Depends(get_db)):
+    """Rota para buscar o pedido de evidencia pelo id"""
+    return request_for_evidence_crud.get_request_for_evidence(id=id, db=db)
+
+@router.post("/request_for_evidence/", response_model=schemas.RequestForEvidence)
+def create_request_for_evidence(request_for_evidence: schemas.RequestForEvidenceCreate, db: Session = Depends(get_db)):
+    """Rota para criar um pedido de evidencia pelo id"""
+    return request_for_evidence_crud.create_request_for_evidence(db=db, request_for_evidence=request_for_evidence)
+
+@router.put("/request_for_evidence/", response_model=schemas.RequestForEvidence)
+def update_request_for_evidence(request_for_evidence: schemas.RequestForEvidence , db: Session = Depends(get_db)):
+    """Rota para alterar o pedido de evidencia pelo id"""
+    return request_for_evidence_crud.update_request_for_evidence(request_for_evidence=request_for_evidence, db=db)
+
+@router.delete("/request_for_evidence/{id}")
+def delete_request_for_evidence(id: int, db: Session = Depends(get_db)):
+    """Rota para deletar o pedido de evidencia pelo id"""
+    return request_for_evidence_crud.delete_request_for_evidence(id=id, db=db)
