+++ src/projeto-backend/routers/user.py
@@ -1,0 +1,75 @@
+from fastapi import APIRouter, Depends, HTTPException
+from database import schemas
+from sqlalchemy.orm import Session
+from models import user_crud
+from database.database import get_db
+from typing import List
+
+router = APIRouter(
+    tags=['Users']
+)
+
+
+## Rotas que trabalham o usuario
+
+
+
+@router.get("/users/{id}", response_model=schemas.UserGet)
+def get_user(id: int, db: Session = Depends(get_db)):
+    """ Você coloca o id do usuario e te retorna um json contendo todos seus dados e processos e etapas
+     que ele está responsável e caso ele não exista lança um exceção http falando que ele não existe"""
+    user = user_crud.get_user(id=id, db=db)
+    if not user:
+        raise HTTPException(status_code=404, detail="Não há Usuarios cadastrado")
+    
+    return user
+
+
+
+
+@router.get("/users/", response_model=List[schemas.UserGet])
+def get_all_users(db: Session = Depends(get_db)):
+    """ Você coloca o id do usuario e te retorna um json contendo todos seus dados e processos e etapas
+     que ele está responsável e caso ele não exista lança um exceção http falando que ele não existe"""
+    user = user_crud.get_all_user(db=db)
+    if not user:
+        raise HTTPException(status_code=404, detail="Usuario não existe")
+    
+    return user
+    
+
+
+
+
+@router.post("/users/", response_model=schemas.User)
+def create_user(user: schemas.UserCreate, db: Session = Depends(get_db)):
+    """Você posta no body os dados do usuario seguindo o modelo da classe UserCreate e se aquele email 
+    não tiver sido registrado ainda será adicionado o usuario o banco, caso ele exista lança uma http 
+    exception falando que aquele email ja está registrado"""
+    db_user = user_crud.get_user_by_email(db, email=user.email)
+    if db_user:
+        raise HTTPException(status_code=400, detail="Email already registered")
+    return user_crud.create_user(db=db, user=user)
+
+
+
+
+
+@router.put("/users/", response_model=schemas.User)
+def update_user(user: schemas.User , db: Session = Depends(get_db)):
+    """Você recebe o usuario que você quer alterar e o altera caso ele exista no banco, caso ele não 
+    exista lança uma exceção dizendo que usuario não foi encontrado"""
+    user = user_crud.update_user(user=user, db=db)
+    if not user:
+        raise HTTPException(status_code=404, detail="Usuario não encotrado")
+    
+    return user
+
+
+
+
+@router.delete("/users/{id}")
+def delete_user(id: int, db: Session = Depends(get_db)):
+    """Recebe o id do usuario e tenta deletar ele no banco caso ele exista ira deleta-lo e retorna ele 
+    mostrando seus dados, caso não tenho retornara None"""
+    return user_crud.delete_user(id=id, db=db)
